scaling updates

+++ components/GroupListHeader.js
@@ -22,7 +22,10 @@ function GroupListHeader (props) {
     FileSystem.readDirectoryAsync(FileSystem.documentDirectory).then(
       contents => {
         Object.keys(props.activeDatabase.sources).forEach(source => {
-          if (!contents.includes(`${props.languageID***REMOVED***-${source***REMOVED***.mp3`))
+          if (
+            !contents.includes(`${props.languageID***REMOVED***-${source***REMOVED***.mp3`) &&
+            !contents.includes(`${props.languageID***REMOVED***-${source***REMOVED***.png`)
+          )
             deleteLanguageInstance()
         ***REMOVED***)
       ***REMOVED***
@@ -135,8 +138,8 @@ const styles = StyleSheet.create({
   languageHeaderContainer: {
     alignItems: 'center',
     width: '100%',
-    // height: 40 * scaleMultiplier,
-    aspectRatio: 8.7,
+    height: 40 * scaleMultiplier,
+    // aspectRatio: 8.7,
     backgroundColor: colors.aquaHaze
   ***REMOVED***,
   languageLogo: {
